refactor: 패키지 및 앱아이디 수정 expbox.app -> expbox
모듈에 클린 아키텍쳐의 형태로 패키지를 구성하려는데,
app 이란 엔드포인트는 어울리지 않아 개발 초기인 지금 바꾸는 것이 좋다.
