+++ app/src/androidTest/java/com/dgparkcode/expbox/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.dgparkcode.expbox.app
+package com.dgparkcode.expbox
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
